requirements.dash.txt
dash>=2.17.0
dash-bootstrap-components>=1.6.0
plotly>=5.22.0
requests>=2.31.0
python-dotenv>=1.0.0